backend/requirements.txt
pymongo[srv]
fastapi
uvicorn
pydantic
dotenv
passlib
httpx
pytest
bcrypt
python-jose[cryptography]
python-multipart
jinja2
authlib
fastapi-mail
starlette[full]
itsdangerous